Encapsulate common sdist actions into a cmdclass.
The pattern of running generate_authors and write_git_changelog
at sdist time is (and should be) repeated in all project's setup.py.
Instead of copying the creation of the two classes and injection
of them into a cmdclass dict, we can do it here and have it
be consistent.

Related to bug 986462.

Change-Id: I0b7b44d58685551482f874b2a6091dec29e6bed6